Extract writingSystem from types/project.ts
